backend/pyproject.toml
[tool.poetry]
name = "todo-evolution-backend"
version = "0.1.0"
description = "Backend for the Todo Evolution project"
authors = ["Metachemist <metachemist@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlmodel = "^0.0.16"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
better-exceptions = "^0.3.3"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
bcrypt = "^4.0.1"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.25.0"
httpx = "^0.27.2"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.main:main"